Fix minor doc typos
Two documentation changes:

- remove period from middle of sentence
- make link formatting consistent

More specifically, the documentation for every Stan inference algorithm on this page, follows the form of "Stan's <some algorithm>", with a link to the algorithm itself.

However, the optimization algorithm is the only type of inference algorithm that includes the word "Stan" as part of the link. No other inference algorithm does this.